refactor: rename plugin.js to index.js
Aligns the package layout with the npm convention (and with
nearform/fastify-slow-down#625) so consumers can simply
`require('fastify-casbin-rest')` against the package root
without depending on a custom file name. The matching types
file moves to types/index.d.ts so declaration lookup and
runtime resolution stay co-located, and the type test
imports switch to './index.js'.
